+++ subprojects/github-publish/src/main/kotlin/github/publish/GitHubPublicationFactory.kt
@@ -1,5 +1,7 @@
-package com.nisecoder.gradle.plugin.github
+package com.nisecoder.gradle.plugin.github.publish
 
+import com.nisecoder.gradle.plugin.github.publish.DefaultGitHubPublication
+import com.nisecoder.gradle.plugin.github.publish.GitHubPublication
 import org.gradle.api.NamedDomainObjectFactory
 import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
